_channel_norm_2d module
