+++ neuralcompression/models/_hific/_channel_norm_2d.py
@@ -1,0 +1,38 @@
+import torch
+import torch.nn
+
+
+def _channel_norm_2d(input_channels, momentum=0.1, affine=True, track_running_stats=False, **kwargs):
+    return ChannelNorm2D(
+        input_channels,
+        momentum=momentum,
+        affine=affine,
+        track_running_stats=track_running_stats
+    )
+
+
+class ChannelNorm2D(torch.nn.Module):
+    def __init__(self, input_channels, momentum=0.1, epsilon=1e-3, affine=True, **kwargs):
+        super(ChannelNorm2D, self).__init__()
+
+        self.momentum = momentum
+
+        self.epsilon = epsilon
+
+        self.affine = affine
+
+        if affine is True:
+            self.gamma = torch.nn.Parameter(torch.ones(1, input_channels, 1, 1))
+            self.beta = torch.nn.Parameter(torch.zeros(1, input_channels, 1, 1))
+
+    def forward(self, x):
+        mean = torch.mean(x, dim=1, keepdim=True)
+
+        variance = torch.var(x, dim=1, keepdim=True)
+
+        x_normed = (x - mean) * torch.rsqrt(variance + self.epsilon)
+
+        if self.affine is True:
+            x_normed = self.gamma * x_normed + self.beta
+
+        return x_normed
